# docs/openapi-3.0.yaml
openapi: 3.0.0
info:
  contact:
    email: support@volumeviz.io
    name: API Support
    url: https://github.com/mantonx/volumeviz/issues
  description: >-
    Docker volume monitoring API with comprehensive volume discovery, size
    calculation, and container attachment tracking. Focus on user-mounted
    volumes only.


    ## Volume-First Approach

    - Automatic discovery and filtering of user-mounted volumes

    - Excludes Docker infrastructure volumes (container filesystems, tmp volumes)

    - Real-time volume usage monitoring and size calculation

    - Container attachment tracking for each volume


    ## Features

    - Multi-method volume size calculation (du, find, stat)

    - Asynchronous scanning with progress tracking for large volumes

    - Real-time WebSocket progress updates during scan operations

    - High-performance caching with TTL-based invalidation

    - Comprehensive Prometheus metrics integration

    - Circuit breaker patterns for resilience


    ## WebSocket API

    WebSocket endpoint: `ws://localhost:8080/ws`


    ### Scan Progress Events

    Subscribe to real-time scan progress updates:

    ```json

    {

    "type": "subscribe",

    "event": "scan_progress",

    "data": {

    "volume_id": "volume_id_here",

    "scan_id": "scan_id_here"

    }

    }

    ```


    Progress updates are broadcast as:

    ```json

    {

    "type": "scan_progress_update",

    "data": {

    "scan_id": "scan_id",

...

...

...

    "phases": [

    {

...

...

...
paths:
...
  /api/v1/mounts:
    get:
      description: Returns paginated list of Docker mount catalog entries with
        advanced filtering, search, and sorting capabilities
      parameters:
        - description: "Page number (default: 1)"
          example: 1
          in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - description: "Page size (default: 25)"
          example: 25
          in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - description: "Sort by: mount_type, compose_project, last_seen, container_count"
          example: '"mount_type"'
          in: query
          name: sort
          schema:
            type: string
        - description: Search query for general text search across mount fields
          example: '"myapp"'
          in: query
          name: q
          schema:
            type: string
        - description: Filter by mount ID (partial match)
          example: '"vol_"'
          in: query
          name: mount_id
          schema:
            type: string
        - description: Filter by volume name (partial match)
          example: '"data"'
          in: query
          name: volume_name
          schema:
            type: string
        - description: Filter by Compose project (partial match)
          example: '"myproject"'
          in: query
          name: compose_project
          schema:
            type: string
        - description: Filter by Compose service (partial match)
          example: '"web"'
          in: query
          name: compose_service
          schema:
            type: string
        - description: Filter by mount type
          example: '"volume"'
          in: query
          name: type
          schema:
            type: string
            enum:
              - volume
              - bind
              - tmpfs
        - description: Filter by status
          example: '"active"'
          in: query
          name: status
          schema:
            type: string
            enum:
              - orphaned
              - active
        - description: Filter by orphaned status
          example: '"false"'
          in: query
          name: is_orphaned
          schema:
            type: string
            enum:
              - "true"
              - "false"
        - description: Filter by tracking status
          example: '"true"'
          in: query
          name: is_tracked
          schema:
            type: string
            enum:
              - "true"
              - "false"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
      summary: List mount catalog entries
      tags:
        - docker-mounts
  "/api/v1/mounts/{id}":
    get:
      description: Returns detailed metadata for a specific mount including volume
        info, compose metadata, and tracking status
      parameters:
        - description: Mount ID
          example: '"vol_abc123"'
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_mounts.MountCatalogResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
      summary: Get mount details
      tags:
        - docker-mounts
  "/api/v1/mounts/{mount_id}/tracking":
    put:
      description: Enable or disable tracking for a specific mount. When tracking is
        enabled, the mount will be included in volume scans and analysis.
      parameters:
        - description: Mount ID
          example: '"vol_abc123"'
          in: path
          name: mount_id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        description: Tracking status
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_mounts.MountCatalogResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                type: object
      summary: Update mount tracking status
      tags:
        - docker-mounts
  /api/v1/mounts/discover:
    post:
      description: Triggers discovery and cataloging of Docker mounts from the Docker
        Engine. Scans all running containers and volumes to build/update the
        mount catalog.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_v1_mounts.DiscoverMountsRequest"
        description: Discovery options
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_mounts.DiscoverMountsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
      summary: Trigger mount discovery
      tags:
        - docker-mounts
  /api/v1/mounts/summary:
    get:
      description: Returns summary statistics for the Docker mount catalog including
        counts by type, orphaned status, and tracking status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_mounts.MountCatalogSummaryResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
      summary: Get mount catalog summary
      tags:
        - docker-mounts
...
servers:
...
components:
  schemas:
...
    internal_api_v1_mounts.DiscoverMountsRequest:
      properties:
        force:
          example: false
          type: boolean
      type: object
    internal_api_v1_mounts.DiscoverMountsResponse:
      properties:
        message:
          example: Mount discovery completed successfully
          type: string
        triggered:
          example: true
          type: boolean
      type: object
    internal_api_v1_mounts.MountCatalogResponse:
      properties:
        compose_project:
          example: myapp
          type: string
        compose_services:
          example:
            - web
            - api
          items:
            type: string
          type: array
        compose_version:
          example: "3.8"
          type: string
        container_count:
          example: 2
          type: integer
        created_at:
          example: 2024-01-01T12:00:00Z
          format: date-time
          type: string
        discovery_source:
          example: docker_engine
          type: string
        first_discovered_at:
          example: 2024-01-01T12:00:00Z
          format: date-time
          type: string
        id:
          example: 1
          type: integer
        is_orphaned:
          example: false
          type: boolean
        is_tracked:
          example: true
          type: boolean
        last_seen_at:
          example: 2024-01-01T14:30:00Z
          format: date-time
          type: string
        mount_id:
          example: vol_abc123
          type: string
        mount_type:
          enum:
            - volume
            - bind
            - tmpfs
          example: volume
          type: string
        source_path:
          example: /var/lib/docker/volumes/my-app-data/_data
          type: string
        tracking_disabled_at:
          format: date-time
          type: string
        tracking_enabled_at:
          example: 2024-01-01T12:00:00Z
          format: date-time
          type: string
        updated_at:
          example: 2024-01-01T14:30:00Z
          format: date-time
          type: string
        volume_driver:
          example: local
          type: string
        volume_name:
          example: my-app-data
          type: string
        volume_scope:
          example: local
          type: string
      type: object
    internal_api_v1_mounts.MountCatalogSummaryResponse:
      properties:
        bind_mounts:
          example: 8
          type: integer
        compose_projects:
          example: 5
          type: integer
        orphaned_mounts:
          example: 3
          type: integer
        tmpfs_mounts:
          example: 2
          type: integer
        total_mounts:
          example: 25
          type: integer
        tracked_mounts:
          example: 20
          type: integer
        volume_mounts:
          example: 15
          type: integer
      type: object
...